refactor(app): extract useShellShortcuts from session-route
Command palette / session search / terminal open-state plus the global
Cmd-N/K/J/Shift-F keydown handler (useEffectEvent) move verbatim to
shell/use-shell-shortcuts.ts. The hook reads canCreateTask/onCreateTask at
event time exactly as before; the call sits after
handleCreateTaskInWorkspace so nothing is referenced before declaration,
with the render watchdog relocated alongside to keep its panel-state
attribution fields.

session-route.tsx: 2,815 -> 2,777 lines.

Tested: pnpm typecheck clean; bun test 58 pass; pnpm build clean;
pnpm test:e2e 23/23 assertions ok.
